requirements.txt
paho-mqtt>=1.6.1
python-dotenv>=0.19.0
requests>=2.26.0